@@ -18,40 +18,46 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  firebase_core: ^1.1.0
-  firebase_analytics: ^8.0.3
-  firebase_auth: ^1.1.3
-  cloud_firestore: ^2.0.0
-  firebase_crashlytics: ^2.0.2
-  firebase_storage: ^8.0.5
-  permission_handler: ^7.1.0
+#  firebase_core_web: ^1.1.0
+  firebase_core: ^1.2.0
+  firebase_analytics: ^8.1.0
+  firebase_auth: ^1.2.0
+  cloud_firestore: ^2.2.0
+  firebase_messaging: ^10.0.0
+  flutter_local_notifications: ^6.0.0
+  firebase_messaging_platform_interface: ^3.0.0
+  overlay_support: ^1.2.1
+  firebase_crashlytics: ^2.0.4
+  firebase_storage: ^8.1.0
+  permission_handler: ^8.0.0+2
   intl: ^0.17.0
-  sqflite_sqlcipher: ^1.1.4
-  http: ^0.13.0
-  flutter_spinkit: ^4.0.0
-  path_provider: ^2.0.1
-  hexcolor: ^2.0.0
-  shared_preferences: ^2.0.3
-  sms_otp_auto_verify: ^1.2.2
-  json_annotation: ^4.0.0
-  webview_flutter: ^2.0.1
-  modal_bottom_sheet: ^1.0.0+1
-  group_list_view: ^1.1.0
+  sqflite_sqlcipher: ^2.0.0
+  http: ^0.13.3
+  flutter_spinkit: ^5.0.0
+  path_provider: ^2.0.2
+  hexcolor: ^2.0.3
+  shared_preferences: ^2.0.6
+#  sms_otp_auto_verify: ^1.2.2
+  json_annotation: ^4.0.1
+  webview_flutter: ^2.0.8
+  modal_bottom_sheet: ^2.0.0
+  group_list_view: ^1.1.1
   carousel_slider: ^3.0.0
-  rounded_loading_button: ^1.0.0
-  image_picker: ^0.7.4
+  rounded_loading_button: ^2.0.4
+  image_picker: ^0.7.5+3
   uuid: ^3.0.4
-  file_picker: ^3.0.1
+  file_picker: ^3.0.2+2
   audioplayers: ^0.19.0
-
+  universal_platform: ^0.1.3
+  universal_html: ^2.0.8
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.3
 
 dev_dependencies:
   flutter_test:
